// Table Dialog Example/wf-app/src/app/models/process-element.ts
export interface ProcessElement {
    status: string;
    processKeyNumber : number;
    alphaNumeric10: string;
    alphaNumeric33: string;
    alphaNumeric57: string;
    volumeCount : number;
    comments: string;
    name: string;
    position: number;
    weight: number;
    symbol: string;
  }